+++ apps/back-office/src/components/base-layout/base-layout.type.ts
@@ -1,0 +1,10 @@
+export interface LayoutRoute {
+  pathname: string;
+  title: string;
+}
+
+export interface LayoutOption {
+  pathname: string;
+  hasHeader?: boolean;
+  hasSidebar?: boolean;
+}
