+++ apps/back-office/src/components/base-layout/hooks/index.ts
@@ -1,0 +1,3 @@
+import useLayoutPathname from "./use-layout-pathname";
+
+export default useLayoutPathname;
