+++ apps/back-office/src/components/base-layout/index.ts
@@ -1,0 +1,3 @@
+import BaseLayout from "./base-layout";
+
+export default BaseLayout;
